+++ designer/designer-ejb/src/java/InterfacePersistencia/PersistenciaSolucionesNodosInterface.java
@@ -146,7 +146,7 @@ public interface PersistenciaSolucionesNodosInterface {
 
     public List<SolucionesNodos> buscarSolucionesNodosParaParametroContable(EntityManager em, Date fechaInicial, Date fechaFinal);
 
-    public Long activos(EntityManager em, BigInteger secuencia);
+    public BigDecimal activos(EntityManager em, BigInteger secuencia);
 
     public List<SolucionesNodos> buscarSolucionesNodosParaParametroContable_SAP(EntityManager em, Date fechaInicial, Date fechaFinal);
